Download config from server and perform analysis

+++ contracd/src/exposureconfiguration.h
@@ -8,11 +8,11 @@ class ExposureConfiguration : public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(quint8 minimumRiskScore READ minimumRiskScore WRITE setMinimumRiskScore NOTIFY minimumRiskScoreChanged)
-    Q_PROPERTY(QList<quint32> attenuationScores READ attenuationScores WRITE setAttenuationScores NOTIFY attenuationScoresChanged)
-    Q_PROPERTY(QList<quint32> daysSinceLastExposureScores READ daysSinceLastExposureScores WRITE setDaysSinceLastExposureScores NOTIFY daysSinceLastExposureScoresChanged)
-    Q_PROPERTY(QList<quint32> durationScores READ durationScores WRITE setDurationScores NOTIFY durationScoresChanged)
-    Q_PROPERTY(QList<quint32> transmissionRiskScores READ transmissionRiskScores WRITE setTransmissionRiskScores NOTIFY transmissionRiskScoresChanged)
+    Q_PROPERTY(qint32 minimumRiskScore READ minimumRiskScore WRITE setMinimumRiskScore NOTIFY minimumRiskScoreChanged)
+    Q_PROPERTY(QList<qint32> attenuationScores READ attenuationScores WRITE setAttenuationScores NOTIFY attenuationScoresChanged)
+    Q_PROPERTY(QList<qint32> daysSinceLastExposureScores READ daysSinceLastExposureScores WRITE setDaysSinceLastExposureScores NOTIFY daysSinceLastExposureScoresChanged)
+    Q_PROPERTY(QList<qint32> durationScores READ durationScores WRITE setDurationScores NOTIFY durationScoresChanged)
+    Q_PROPERTY(QList<qint32> transmissionRiskScores READ transmissionRiskScores WRITE setTransmissionRiskScores NOTIFY transmissionRiskScoresChanged)
 
     Q_PROPERTY(double attenuationWeight READ attenuationWeight WRITE setAttenuationWeight NOTIFY attenuationWeightChanged)
     Q_PROPERTY(double daysSinceLastExposureWeight READ daysSinceLastExposureWeight WRITE setDaysSinceLastExposureWeight NOTIFY daysSinceLastExposureWeightChanged)
@@ -25,22 +25,22 @@ public:
     ExposureConfiguration(ExposureConfiguration const &exposureConfiguration);
     ExposureConfiguration& operator=( const ExposureConfiguration &other);
 
-    quint8 minimumRiskScore() const;
-    QList<quint32> attenuationScores() const;
-    QList<quint32> daysSinceLastExposureScores() const;
-    QList<quint32> durationScores() const;
-    QList<quint32> transmissionRiskScores() const;
+    qint32 minimumRiskScore() const;
+    QList<qint32> attenuationScores() const;
+    QList<qint32> daysSinceLastExposureScores() const;
+    QList<qint32> durationScores() const;
+    QList<qint32> transmissionRiskScores() const;
     double attenuationWeight() const;
     double daysSinceLastExposureWeight() const;
     double durationWeight() const;
     double transmissionRiskWeight() const;
     QList<qint32> durationAtAttenuationThresholds() const;
 
-    void setMinimumRiskScore(quint8 minimumRiskScore);
-    void setAttenuationScores(QList<quint32> attenuationScores);
-    void setDaysSinceLastExposureScores(QList<quint32> daysSinceLastExposureScores);
-    void setDurationScores(QList<quint32> durationScores);
-    void setTransmissionRiskScores(QList<quint32> transmissionRiskScores);
+    void setMinimumRiskScore(qint32 minimumRiskScore);
+    void setAttenuationScores(QList<qint32> attenuationScores);
+    void setDaysSinceLastExposureScores(QList<qint32> daysSinceLastExposureScores);
+    void setDurationScores(QList<qint32> durationScores);
+    void setTransmissionRiskScores(QList<qint32> transmissionRiskScores);
     void setAttenuationWeight(double attenuationWeight);
     void setDaysSinceLastExposureWeight(double daysSinceLastExposureWeight);
     void setDurationWeight(double durationWeight);
@@ -60,11 +60,11 @@ signals:
     void durationAtAttenuationThresholdsChanged();
 
 private:
-    quint8 m_minimumRiskScore;
-    QList<quint32> m_attenuationScores;
-    QList<quint32> m_daysSinceLastExposureScores;
-    QList<quint32> m_durationScores;
-    QList<quint32> m_transmissionRiskScores;
+    qint32 m_minimumRiskScore;
+    QList<qint32> m_attenuationScores;
+    QList<qint32> m_daysSinceLastExposureScores;
+    QList<qint32> m_durationScores;
+    QList<qint32> m_transmissionRiskScores;
 
     double m_attenuationWeight;
     double m_daysSinceLastExposureWeight;
